pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "transpiler-mate"
dynamic = ["version"]
description = 'Schema.org 2 metadata formats transpiler'

readme = "README.md"
requires-python = ">=3.8"
license = "Apache-2.0"
keywords = []
authors = [
  { name = "Fabrice Brito", email = "info@terradue.com" },
  { name = "Simone Tripodi", email = "info@terradue.com"}
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "pydantic",
  "email-validator",
  "ruamel.yaml>=0.18.16",
  "yaml-ld",
  "click>=8.3.0",
  "loguru",
  "invenio-rest-api-client",
  "giturlparse",
  "semver"
]

[tool.hatch.metadata]
allow-direct-references = true

[project.urls]
Documentation = "https://github.com/Terradue/transpiler-mate#readme"
Issues = "https://github.com/Terradue/transpiler-mate/issues"
Source = "https://github.com/Terradue/transpiler-mate"

[project.scripts]
transpiler-mate = "transpiler_mate.cli:main"

[tool.hatch.version]
path = "src/transpiler_mate/__about__.py"

[tool.hatch.envs.types]
extra-dependencies = [
  "mypy>=1.0.0",
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/transpiler_mate tests}"

[tool.coverage.run]
source_pkgs = ["transpiler_mate", "tests"]
branch = true
parallel = true
omit = [
  "src/transpiler_mate/__about__.py",
]

[tool.coverage.paths]
transpiler_mate = ["src/transpiler_mate", "*/transpiler_mate/src/transpiler_mate"]
tests = ["tests", "*/transpiler_mate/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]

[tool.hatch.envs.test]
skip-install = false

dependencies = [
  "pydantic",
  "email-validator",
  "ruamel.yaml>=0.18.16",
  "yaml-ld",
  "click>=8.3.0",
  "loguru",
  "invenio-rest-api-client",
  "giturlparse",
  "semver"
]

[tool.hatch.envs.test.scripts]
test = "hatch run nose2"
testv = "hatch run nose2 --verbose"
cov = ["coverage run --source=src/transpiler_mate -m nose2", "coverage report"]

[[tool.hatch.envs.test.matrix]]
python = ["3.9", "3.10", "3.11", "3.12", "3.13"]
